Moves independent test utils functions into test_process_list_utils
Fixes the imports to point to the new module
Small formatting changes from format on save
