+++ savu/test/travis/framework_tests/plugin_coverage_test.py
@@ -26,6 +26,7 @@ import os
 
 from savu.data.plugin_list import PluginList
 import savu.test.test_utils as tu
+import savu.test.test_process_list_utils as tplu
 
 
 class PluginCoverageTest(unittest.TestCase):
@@ -37,13 +38,13 @@ class PluginCoverageTest(unittest.TestCase):
         # lists all .nxs process lists used in the tests, and all plugins
         # directly called in the tests
         [nxs_in_tests, plugins_in_tests] = \
-            tu.get_process_list(savu_base_path + '/savu/test')
+            tplu.get_process_list(savu_base_path + '/savu/test')
 
         # remove data files from the list
         data_list = self.get_data_list(savu_base_path + '/test_data/data')
         nxs_in_tests = list(set(nxs_in_tests).difference(set(data_list)))
-        
-        # since now some nxs files are in the subfolders, we need to make sure 
+
+        # since now some nxs files are in the subfolders, we need to make sure
         # that we extract the basename only (without the "head")
         nxs_in_tests_mod = []
         for nxs in nxs_in_tests:
@@ -56,7 +57,7 @@ class PluginCoverageTest(unittest.TestCase):
 
         # list all test process lists available in test_process_lists folder
         test_process_path = savu_base_path + 'test_data/test_process_lists'
-        self.nxs_avail = tu.get_test_process_list(test_process_path)
+        self.nxs_avail = tplu.get_test_process_list(test_process_path)
 
         # list the .nxs found in tests that are located in the
         # test_process_lists folder
@@ -65,13 +66,13 @@ class PluginCoverageTest(unittest.TestCase):
 
         # which test process lists were not in the test_process_lists folder
         nxs_unused = list(set(nxs_in_tests).difference(set(self.nxs_avail)))
-        print ("===============================================================")
-        print ("\nThese .nxs test files were found inside the tests, but are "
-               "not available in the test_process_lists folder:\n")
+        print("===============================================================")
+        print("\nThese .nxs test files were found inside the tests, but are "
+              "not available in the test_process_lists folder:\n")
         for nxs in nxs_unused:
             print (nxs)
         print ("===============================================================")
-        
+
         # get all plugins listed in self.nxs_used process lists
         tested_plugin_list = \
             self.get_test_plugin_list(self.nxs_used, test_process_path)
@@ -79,7 +80,7 @@ class PluginCoverageTest(unittest.TestCase):
 
         # list all plugins
         plugin_list = self.get_plugin_list(savu_base_path + '/savu/plugins')
-        
+
         print ("===============================================================")
         print ("\nThe following plugins are not covered by the tests:\n")
         uncovered = list(set(plugin_list).difference(set(tested_plugin_list)))
@@ -124,8 +125,7 @@ class PluginCoverageTest(unittest.TestCase):
                         plugin_names.append(p + '.py')
                 except ImportError as e:
                     print("Failed to run test as libraries not available (%s),"
-                          % (e) + " passing test")
-                    pass
+                           % (e) + " passing test")
         return list(set(plugin_names))
 
     def add_plugin(self, plugin_id):
@@ -159,6 +159,7 @@ class PluginCoverageTest(unittest.TestCase):
         mod2class = ''.join(x.capitalize() for x in module_name.split('_'))
         clazz = getattr(mod, mod2class.split('.')[-1])()
         return clazz
+
 
 if __name__ == "__main__":
     unittest.main()
